#6371: wrap through UncheckedIOException, keep the smells count unchanged
